commit nabila mercredi 15:50

+++ src/main/java/org/formation/model/Compte.java
@@ -21,7 +21,7 @@ public abstract class Compte {
 	
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
-	private int id;
+	private long id;
 	private double montant;
 	private LocalDate dateOuverture;
 	
@@ -40,10 +40,10 @@ public abstract class Compte {
 	public void setDateOuverture(LocalDate dateOuverture) {
 		this.dateOuverture = dateOuverture;
 	}
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public Compte(double montant, LocalDate dateOuverture) {
@@ -54,6 +54,7 @@ public abstract class Compte {
 	public Compte() {
 		super();
 	}
+
